+++ src/aux-server/aux-web/aux-player/MenuBot/MenuBot.vue
@@ -5,7 +5,11 @@
         :style="{ 'background-color': backgroundColor }"
         @click="click()"
     >
-        <div class="menu-bot-text" v-show="label" :style="{ color: labelColor }">
+        <div
+            class="menu-bot-text"
+            v-show="label"
+            :style="{ color: labelColor, 'text-align': labelAlign }"
+        >
             {{ label }}
         </div>
     </md-list-item>